# api/openapi-v1.json
{
  "components": {
    "schemas": {
      "AnalysisConclusion": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "counterEvidence": {
            "items": {
              "type": "string"
            },
            "title": "Counterevidence",
            "type": "array"
          },
          "drawingRefs": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Drawingrefs",
            "type": "array"
          },
          "figureRefs": {
            "items": {
              "type": "string"
            },
            "title": "Figurerefs",
            "type": "array"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "reasoning": {
            "minLength": 1,
            "title": "Reasoning",
            "type": "string"
          },
          "strength": {
            "enum": [
              "strong",
              "moderate",
              "weak",
              "unclear"
            ],
            "title": "Strength",
            "type": "string"
          },
          "timeAnchor": {
            "minLength": 1,
            "title": "Timeanchor",
            "type": "string"
          },
          "title": {
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "verdict": {
            "minLength": 1,
            "title": "Verdict",
            "type": "string"
          }
        },
        "required": [
          "id",
          "timeAnchor",
          "title",
          "verdict",
          "strength",
          "reasoning",
          "counterEvidence",
          "drawingRefs",
          "figureRefs",
          "confidence"
        ],
        "title": "AnalysisConclusion",
        "type": "object"
      },
      "AnalysisContextV13": {
        "additionalProperties": false,
        "properties": {
          "capturedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capturedat"
          },
          "dataSources": {
            "items": {
              "enum": [
                "screenshot",
                "exchange_api",
                "calculated",
                "user_context"
              ],
              "type": "string"
            },
            "title": "Datasources",
            "type": "array"
          },
          "instrument": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument"
          },
          "latestCandleClosed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestcandleclosed"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "timeframes": {
            "items": {
              "$ref": "#/components/schemas/TimeframeRole"
            },
            "maxItems": 3,
            "title": "Timeframes",
            "type": "array"
          },
          "venue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue"
          }
        },
        "title": "AnalysisContextV13",
        "type": "object"
      },
      "AnalysisProgressEvent": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "enum": [
              "preparing",
              "reading_chart",
              "reviewing_clues",
              "checking_signals",
              "preparing_result"
            ],
            "title": "Code",
            "type": "string"
          },
          "createdAt": {
            "title": "Createdat",
            "type": "string"
          }
        },
        "required": [
          "code",
          "createdAt"
        ],
        "title": "AnalysisProgressEvent",
        "type": "object"
      },
      "AnalysisReport": {
        "additionalProperties": false,
        "properties": {
          "analysisContext": {
            "$ref": "#/components/schemas/AnalysisContextV13"
          },
          "bearishEvidence": {
            "items": {
              "type": "string"
            },
            "title": "Bearishevidence",
            "type": "array"
          },
          "bullishEvidence": {
            "items": {
              "type": "string"
            },
            "title": "Bullishevidence",
            "type": "array"
          },
          "conclusions": {
            "items": {
              "$ref": "#/components/schemas/AnalysisConclusion"
            },
            "maxItems": 6,
            "title": "Conclusions",
            "type": "array"
          },
          "conflicts": {
            "items": {
              "type": "string"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "decision": {
            "$ref": "#/components/schemas/Decision"
          },
          "dominantBias": {
            "enum": [
              "bullish",
              "bearish",
              "neutral",
              "unclear"
            ],
            "title": "Dominantbias",
            "type": "string"
          },
          "drawings": {
            "items": {
              "$ref": "#/components/schemas/DrawingInstruction"
            },
            "maxItems": 24,
            "title": "Drawings",
            "type": "array"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceReferenceV13"
            },
            "title": "Evidence",
            "type": "array"
          },
          "imageQuality": {
            "$ref": "#/components/schemas/ImageQuality"
          },
          "indicatorReadings": {
            "items": {
              "$ref": "#/components/schemas/IndicatorReading"
            },
            "maxItems": 8,
            "title": "Indicatorreadings",
            "type": "array"
          },
          "insights": {
            "items": {
              "$ref": "#/components/schemas/Insight"
            },
            "maxItems": 6,
            "title": "Insights",
            "type": "array"
          },
          "keyLevels": {
            "items": {
              "$ref": "#/components/schemas/KeyLevel"
            },
            "maxItems": 8,
            "title": "Keylevels",
            "type": "array"
          },
          "marketReading": {
            "$ref": "#/components/schemas/MarketReading"
          },
          "marketState": {
            "$ref": "#/components/schemas/MarketStateV13"
          },
          "overallConfidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overallconfidence",
            "type": "number"
          },
          "patterns": {
            "items": {
              "$ref": "#/components/schemas/Pattern"
            },
            "maxItems": 3,
            "title": "Patterns",
            "type": "array"
          },
          "positioningEvidence": {
            "items": {
              "$ref": "#/components/schemas/PositioningEvidence"
            },
            "maxItems": 8,
            "title": "Positioningevidence",
            "type": "array"
          },
          "riskNotice": {
            "minLength": 1,
            "title": "Risknotice",
            "type": "string"
          },
          "scenarios": {
            "$ref": "#/components/schemas/Scenarios"
          },
          "schemaVersion": {
            "const": "1.3",
            "title": "Schemaversion",
            "type": "string"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/MarketSegment"
            },
            "maxItems": 8,
            "title": "Segments",
            "type": "array"
          },
          "setupEvaluation": {
            "$ref": "#/components/schemas/SetupEvaluationV13"
          },
          "timeframeAnalyses": {
            "items": {
              "$ref": "#/components/schemas/TimeframeAnalysis"
            },
            "maxItems": 6,
            "title": "Timeframeanalyses",
            "type": "array"
          },
          "tradeSignals": {
            "items": {
              "$ref": "#/components/schemas/TradeSignal"
            },
            "maxItems": 6,
            "title": "Tradesignals",
            "type": "array"
          },
          "volumeAnalysis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VolumeAnalysis"
              },
              {
                "type": "null"
              }
            ]
          },
          "zones": {
            "items": {
              "$ref": "#/components/schemas/DecisionZoneV13"
            },
            "maxItems": 5,
            "title": "Zones",
            "type": "array"
          }
        },
        "required": [
          "imageQuality",
          "marketReading",
          "bullishEvidence",
          "bearishEvidence",
          "conflicts",
          "dominantBias",
          "overallConfidence",
          "decision",
          "insights",
          "keyLevels",
          "patterns",
          "scenarios",
          "drawings",
          "conclusions",
          "segments",
          "indicatorReadings",
          "volumeAnalysis",
          "positioningEvidence",
          "timeframeAnalyses",
          "tradeSignals",
          "riskNotice",
          "schemaVersion",
          "analysisContext",
          "evidence",
          "marketState",
          "zones",
          "setupEvaluation"
        ],
        "title": "AnalysisReport",
        "type": "object"
      },
      "AnalysisTask": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "additionalProperties": true,
            "title": "Context",
            "type": "object"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "errorCode": {
            "anyOf": [
              {
                "enum": [
                  "CV_IMAGE_INVALID",
                  "CV_PROVIDER_TIMEOUT",
                  "CV_PROVIDER_ERROR",
                  "CV_RESPONSE_INVALID",
                  "CV_CANCELLED",
                  "CV_INTERNAL_ERROR"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "progressEvents": {
            "items": {
              "$ref": "#/components/schemas/AnalysisProgressEvent"
            },
            "title": "Progressevents",
            "type": "array"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalysisReport"
              },
              {
                "type": "null"
              }
            ]
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "processing",
              "awaiting_confirmation",
              "cancel_requested",
              "cancelled",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "status",
          "context"
        ],
        "title": "AnalysisTask",
        "type": "object"
      },
      "Body_create_analysis_v1_analyses_post": {
        "properties": {
          "context": {
            "default": "{}",
            "title": "Context",
            "type": "string"
          },
          "image": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Image",
            "type": "array"
          }
        },
        "required": [
          "image"
        ],
        "title": "Body_create_analysis_v1_analyses_post",
        "type": "object"
      },
...
      "Decision": {
        "additionalProperties": false,
        "properties": {
          "direction": {
            "enum": [
              "long",
              "short",
              "wait"
            ],
            "title": "Direction",
            "type": "string"
          },
          "primaryRisk": {
            "minLength": 1,
            "title": "Primaryrisk",
            "type": "string"
          },
          "status": {
            "enum": [
              "waiting_trigger",
              "waiting_confirmation",
              "conditions_met",
              "invalidated"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "direction",
          "status",
          "summary",
          "primaryRisk"
        ],
        "title": "Decision",
        "type": "object"
      },
...
      "DrawingInstruction": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidenceIds": {
            "items": {
              "type": "string"
            },
            "title": "Evidenceids",
            "type": "array"
          },
          "figureId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Figureid"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "label": {
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/DrawingPoint"
            },
            "maxItems": 2,
            "minItems": 1,
            "title": "Points",
            "type": "array"
          },
          "reason": {
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          },
          "renderBounds": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PricePanelBounds"
              },
              {
                "type": "null"
              }
            ]
          },
          "signalDirection": {
            "anyOf": [
              {
                "enum": [
                  "long",
                  "short"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signaldirection"
          },
          "tool": {
            "enum": [
              "support_line",
              "resistance_line",
              "support_zone",
              "resistance_zone",
              "trend_line",
              "breakout_marker",
              "rejection_marker",
              "time_marker",
              "entry_line",
              "stop_line",
              "target_line",
              "note"
            ],
            "title": "Tool",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tool",
          "label",
          "points",
          "reason",
          "evidenceIds",
          "confidence",
          "figureId",
          "signalDirection",
          "renderBounds"
        ],
        "title": "DrawingInstruction",
        "type": "object"
      },
...
      "EvidenceReferenceV13": {
        "additionalProperties": false,
        "properties": {
          "claim": {
            "minLength": 1,
            "title": "Claim",
            "type": "string"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "id": {
            "pattern": "^NE\\d{3}$",
            "title": "Id",
            "type": "string"
          },
          "observationClass": {
            "enum": [
              "direct",
              "estimated",
              "calculated"
            ],
            "title": "Observationclass",
            "type": "string"
          },
          "source": {
            "enum": [
              "screenshot",
              "exchange_api",
              "calculated",
              "user_context"
            ],
            "title": "Source",
            "type": "string"
          },
          "timeAnchor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeanchor"
          },
          "timeframe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeframe"
          }
        },
        "required": [
          "id",
          "claim",
          "source",
          "observationClass",
          "confidence"
        ],
        "title": "EvidenceReferenceV13",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ImageQuality": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "quality": {
            "enum": [
              "high",
              "medium",
              "low"
            ],
            "title": "Quality",
            "type": "string"
          }
        },
        "required": [
          "quality",
          "limitations",
          "confidence"
        ],
        "title": "ImageQuality",
        "type": "object"
      },
      "IndicatorReading": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "drawingIds": {
            "items": {
              "type": "string"
            },
            "title": "Drawingids",
            "type": "array"
          },
          "id": {
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "name": {
            "enum": [
              "RSI",
              "MACD",
              "OTHER"
            ],
            "title": "Name",
            "type": "string"
          },
          "signals": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Signals",
            "type": "array"
          },
          "state": {
            "minLength": 1,
            "title": "State",
            "type": "string"
          },
          "timeAnchor": {
            "minLength": 1,
            "title": "Timeanchor",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "state",
          "signals",
          "timeAnchor",
          "drawingIds",
          "confidence"
        ],
        "title": "IndicatorReading",
        "type": "object"
      },
      "Insight": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence": {
            "minLength": 1,
            "title": "Evidence",
            "type": "string"
          },
          "kind": {
            "enum": [
              "trend",
              "structure",
              "volatility",
              "volume",
              "momentum",
              "indicator"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "minLength": 1,
            "title": "Label",
            "type": "string"
          },
          "value": {
            "minLength": 1,
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "label",
          "value",
          "evidence",
          "confidence"
        ],
        "title": "Insight",
        "type": "object"
      },
...
      "MarketReading": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "structure": {
            "enum": [
              "hh-hl",
              "lh-ll",
              "range",
              "transition",
              "unclear"
            ],
            "title": "Structure",
            "type": "string"
          },
          "trend": {
            "enum": [
              "bullish",
              "bearish",
              "sideways",
              "unclear"
            ],
            "title": "Trend",
            "type": "string"
          }
        },
        "required": [
          "trend",
          "structure",
          "evidence",
          "confidence"
        ],
        "title": "MarketReading",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ChartViz Community API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/analyses": {
      "post": {
        "operationId": "create_analysis_v1_analyses_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_analysis_v1_analyses_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisTask"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Analysis"
      }
    },
    "/v1/analyses/{analysis_id}": {
      "delete": {
        "operationId": "cancel_analysis_v1_analyses__analysis_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisTask"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Analysis"
      },
      "get": {
        "operationId": "get_analysis_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisTask"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Analysis"
      }
    },
...
    "/v1/models": {
      "get": {
        "operationId": "models_v1_models_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "items": {
                      "additionalProperties": {
                        "type": "string"
                      },
                      "type": "object"
                    },
                    "type": "array"
                  },
                  "title": "Response Models V1 Models Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models"
      }
    }
  }
}